Use size_t for template parameters that size a std::array
std::array uses size_t for its size parameter, so size_t is the natural type for template parameters that directly size a std::array or an alias of one.

+++ include/fill_rand.hpp
@@ -89,7 +89,7 @@ fill_rand(T& x)
         std::unreachable();
 }
 
-template <std::unsigned_integral T, unsigned int N>
+template <std::unsigned_integral T, size_t N>
 void
 fill_rand(std::array<T, N>& container)
 {
